refactor
test_stride

failing case

fix the bug with current invalid miss way counted as valid

refactor regression
