+++ bsg_cache/bsg_cache_non_blocking.v
@@ -124,10 +124,16 @@ module bsg_cache_non_blocking
   logic mhu_idle;
   logic mhu_recover;
 
-  logic [addr_width_p-1:0] mhu_evict_addr;
-  logic mhu_evict_v;
-  logic [addr_width_p-1:0] dma_evict_addr;
-  logic dma_evict_v;
+  //logic [addr_width_p-1:0] mhu_evict_addr;
+  //logic mhu_evict_v;
+  //logic [addr_width_p-1:0] dma_evict_addr;
+  //logic dma_evict_v;
+  logic [lg_ways_lp-1:0] curr_mhu_way_id;
+  logic [lg_sets_lp-1:0] curr_mhu_index;
+  logic curr_mhu_v;
+  logic [lg_ways_lp-1:0] curr_dma_way_id;
+  logic [lg_sets_lp-1:0] curr_dma_index;
+  logic curr_dma_v;
 
   bsg_cache_non_blocking_tl_stage #(
     .id_width_p(id_width_p)
@@ -178,11 +184,18 @@ module bsg_cache_non_blocking
     ,.mhu_idle_i(mhu_idle)
     ,.recover_i(mhu_recover)
 
-    ,.mhu_evict_addr_i(mhu_evict_addr)
-    ,.mhu_evict_v_i(mhu_evict_v)
+    //,.mhu_evict_addr_i(mhu_evict_addr)
+    //,.mhu_evict_v_i(mhu_evict_v)
     
-    ,.dma_evict_addr_i(dma_evict_addr)
-    ,.dma_evict_v_i(dma_evict_v)
+    //,.dma_evict_addr_i(dma_evict_addr)
+    //,.dma_evict_v_i(dma_evict_v)
+    ,.curr_dma_way_id_i(curr_dma_way_id)
+    ,.curr_dma_index_i(curr_dma_index)
+    ,.curr_dma_v_i(curr_dma_v)
+
+    ,.curr_mhu_way_id_i(curr_mhu_way_id)
+    ,.curr_mhu_index_i(curr_mhu_index)
+    ,.curr_mhu_v_i(curr_mhu_v)
   );
 
 
@@ -387,8 +400,11 @@ module bsg_cache_non_blocking
     ,.tag_hit_way_i(tag_hit_way_lo)
     ,.tag_hit_found_i(tag_hit_found_lo)
 
-    ,.evict_addr_o(mhu_evict_addr)
-    ,.evict_v_o(mhu_evict_v)
+    //,.evict_addr_o(mhu_evict_addr)
+    //,.evict_v_o(mhu_evict_v)
+    ,.curr_mhu_way_id_o(curr_mhu_way_id)
+    ,.curr_mhu_index_o(curr_mhu_index)
+    ,.curr_mhu_v_o(curr_mhu_v)
 
     ,.miss_fifo_v_i(miss_fifo_v_lo)
     ,.miss_fifo_entry_i(miss_fifo_data_lo)
@@ -431,8 +447,11 @@ module bsg_cache_non_blocking
     ,.pending_o(dma_pending_li)
     ,.ack_i(dma_ack_lo)
 
-    ,.evict_v_o(dma_evict_v)
-    ,.evict_addr_o(dma_evict_addr)
+    //,.evict_v_o(dma_evict_v)
+    //,.evict_addr_o(dma_evict_addr)
+    ,.curr_dma_way_id_o(curr_dma_way_id)
+    ,.curr_dma_index_o(curr_dma_index)
+    ,.curr_dma_v_o(curr_dma_v)
 
     ,.data_mem_pkt_v_o(dma_data_mem_pkt_v_lo)
     ,.data_mem_pkt_o(dma_data_mem_pkt_lo)
