+++ bsg_cache/bsg_cache_non_blocking_tl_stage.v
@@ -82,11 +82,17 @@ module bsg_cache_non_blocking_tl_stage
     , input mhu_idle_i
     , input recover_i
 
-    // evict address
-    , input [addr_width_p-1:0] mhu_evict_addr_i
-    , input mhu_evict_v_i
-    , input [addr_width_p-1:0] dma_evict_addr_i
-    , input dma_evict_v_i
+    //, input [addr_width_p-1:0] mhu_evict_addr_i
+    //, input mhu_evict_v_i
+    //, input [addr_width_p-1:0] dma_evict_addr_i
+    //, input dma_evict_v_i
+    , input [lg_ways_lp-1:0] curr_mhu_way_id_i
+    , input [lg_sets_lp-1:0] curr_mhu_index_i
+    , input curr_mhu_v_i
+
+    , input [lg_ways_lp-1:0] curr_dma_way_id_i
+    , input [lg_sets_lp-1:0] curr_dma_index_i
+    , input curr_dma_v_i
   );
 
 
@@ -252,21 +258,14 @@ module bsg_cache_non_blocking_tl_stage
 
 
   // miss detection logic
-  // when mhu_evict_v_i is high, then the block address matching
-  // mhu_evict_addr_i should be considered as miss. Similarly, with
-  // dma_evict_v_i. This is because the replacement decision has already
-  // been made, and the DMA transaction has already begun or even finished,
-  // but the MHU does not update the tag until it has completed the miss
-  // handling (e.g. going through the miss FIFO to collect all the secondary
-  // misses). 
 
-  logic [addr_width_p-1:0] block_addr_tl;
-  logic mhu_evict_match;
-  logic dma_evict_match;
+  //logic [addr_width_p-1:0] block_addr_tl;
+  //logic mhu_evict_match;
+  //logic dma_evict_match;
 
-  assign block_addr_tl = {addr_tag_tl, addr_index_tl, {block_offset_width_lp{1'b0}}};
-  assign mhu_evict_match = mhu_evict_v_i & (mhu_evict_addr_i == block_addr_tl);
-  assign dma_evict_match = dma_evict_v_i & (dma_evict_addr_i == block_addr_tl);
+  //assign block_addr_tl = {addr_tag_tl, addr_index_tl, {block_offset_width_lp{1'b0}}};
+  wire mhu_miss_match = curr_mhu_v_i & (tag_hit_way == curr_mhu_way_id_i) & (curr_mhu_index_i == addr_index_tl);
+  wire dma_miss_match = curr_dma_v_i & (tag_hit_way == curr_dma_way_id_i) & (curr_dma_index_i == addr_index_tl);
 
 
   // TL stage output logic
@@ -316,7 +315,7 @@ module bsg_cache_non_blocking_tl_stage
   logic mgmt_op_tl;
 
   assign miss_tl = tag_hit_found
-    ? (mhu_evict_match | dma_evict_match)
+    ? (mhu_miss_match | dma_miss_match)
     : 1'b1;
 
   assign ld_st_hit = v_tl_r & (decode_tl_r.ld_op | decode_tl_r.st_op) & ~miss_tl;
