+++ testing/bsg_cache/regression_non_blocking/.gitignore
@@ -1,2 +1,4 @@
 trace.tr
 trace_rom.v
+*.pyc
+out/
