+++ testing/bsg_cache/regression_non_blocking/Makefile
@@ -5,25 +5,20 @@ include ../../../../bsg_cadenv/cadenv.mk
 export BASEJUMP_STL_DIR = $(abspath ../../..)
 
 
-INCDIR = +incdir+$(BASEJUMP_STL_DIR)/bsg_misc/
+INCDIR =  +incdir+$(BASEJUMP_STL_DIR)/bsg_misc/
 INCDIR += +incdir+$(BASEJUMP_STL_DIR)/bsg_cache/
 
-VCS_FLAGS = +v2k -sverilog -timescale=1ps/1ps -O4
+VCS_FLAGS =  -full64 +lint=all,noSVA-UA,noSVA-NSVU,noVCDE -debug_access
+VCS_FLAGS += +v2k -sverilog -timescale=1ps/1ps -O4
+#VCS_FLAGS += -lca -cm line+cond+fsm+tgl+branch+assert
 
-### Options #####
-ifeq ($(WAVE), 1)
-VCS_FLAGS += +vcs+vcdpluson
-endif
+#### Test Suite #####
 
-ifeq ($(LOG), 1)
-VCS_FLAGS += -l vcs.log
-endif
+TEST_SUITE = test_stride test_random test_square test_block test_zorder test_byte \
+						 test_random_flush test_clean_read
 
-ifeq ($(COV), 1)
-VCS_FLAGS += -lca -cm line+cond+fsm+tgl+path+branch+assert
-endif
-#################
 
+#####################
 
 HIGHLIGHT = grep --color -E '^|Fatal|Error|Warning|Implicit wire is used|Too few instance port connections|Port connection width mismatch|Width mismatch'
 
@@ -46,33 +41,35 @@ VCS_DEFINE += +define+YUMI_MAX_DELAY_P=$(YUMI_MAX_DELAY_P)
 
 ########################
 
+SIMV = $(abspath simv)
+WAVE ?= 0
 
-all: sim
+all: $(addsuffix .run, $(TEST_SUITE))
 
-TRACE_TR = trace.tr
-TRACE_GEN_PY = trace_gen.py
-TRACE_ROM = trace_rom
+simv:
+	$(VCS) $(VCS_FLAGS) -f sv.include $(INCDIR) $(VCS_DEFINE) -l vcs.log | $(HIGHLIGHT)
+	
+out/%/trace.tr:
+	mkdir -p out/
+	mkdir -p out/$*
+	python $*.py > $@
 
-$(TRACE_TR): $(TRACE_GEN_PY)
-	python $(TRACE_GEN_PY) > $(TRACE_TR)
+%.run: simv out/%/trace.tr
+	(cd out/$*; $(SIMV) +wave=$(WAVE) -l simv.log)
 
-sim: $(TRACE_TR)
-	$(VCS) -full64 -R +lint=all,noSVA-UA,noSVA-NSVU,noVCDE \
-	  -f sv.include $(INCDIR) \
-		$(VCS_FLAGS) $(VCS_DEFINE) \
-		| $(HIGHLIGHT)
-
-dve:
-	$(DVE) -full64 -vpd vcdplus.vpd &
+%.dve:
+	$(DVE) -full64 -vpd out/$*/vcdplus.vpd &
 
 cov:
 	$(DVE) -full64 -cov -dir simv.vdb &
 
+summary:
+	@$(foreach test, $(TEST_SUITE), grep -H --color -e "BSG_FINISH" -e "BSG_FATAL" out/$(test)/simv.log;)
+
 clean:
 	rm -rf DVEfiles
 	rm -rf csrc
-	rm -rf simv.daidir simv.vdb
-	rm -f ucli.key vcdplus.vpd simv cm.log *.tar.gz vcs.log
-	rm -rf stack.info.*
-	rm -f $(TRACE_TR)
-
+	rm -rf simv.daidir simv.vdb vcs.log
+	rm -f ucli.key vcdplus.vpd simv cm.log *.tar.gz
+	rm -rf stack.info.* *.pyc
+	rm -rf out
