+++ testing/bsg_cache/regression_non_blocking/test_random.py
@@ -1,0 +1,28 @@
+import sys
+import random
+from test_base import *
+
+
+class TestRandom(TestBase):
+
+  def generate(self):
+    # scrub tag and data
+    self.clear_tag()
+    for i in range(self.MAX_ADDR/4):
+      self.send(SW, 4*i)
+
+    # random SW/LW
+    for n in range(200000):
+      addr = random.randint(0, (self.MAX_ADDR/4)-1)*4
+      store_not_load = random.randint(0,1)
+      if store_not_load:
+        self.send(SW, addr)
+      else:
+        self.send(LW, addr)
+
+    self.tg.done()
+          
+#   main()
+if __name__ == "__main__":
+  rd = TestRandom()
+  rd.generate()
